Implement new overview data (#1127)
* * Add to be awarded
* Add fiatValue
* Add open gov count

* In order to compatible with old UI,we keep both V1 and V2 data structure running

* Update
