+++ packages/server/config/config.example.js
@@ -9,6 +9,7 @@ module.exports = {
     dotOutputDbName: "dotreasury-output-dot",
     dotCouncilDbName: "dotreasury-council-dot",
     adminDbName: "dot-treasury-admin",
+    priceDbName: "price",
   },
   server: {
     port: 3213,
