+++ packages/server/src/mongo-price/index.js
@@ -1,0 +1,46 @@
+const { MongoClient } = require("mongodb");
+const config = require("../../config");
+
+const dbName = config.mongo.priceDbName || "price";
+
+const ksmUsdtCollectionName = "ksmUsdt";
+const dotUsdtCollectionName = "dotUsdt";
+
+let client = null;
+let db = null;
+
+const mongoUrl = process.env.MONGO_URL || "mongodb://127.0.0.1:27017";
+let ksmUsdtCol = null;
+let dotUsdtCol = null;
+
+async function initDb() {
+  client = await MongoClient.connect(mongoUrl, {
+    useUnifiedTopology: true,
+  });
+
+  db = client.db(dbName);
+  ksmUsdtCol = db.collection(ksmUsdtCollectionName);
+  dotUsdtCol = db.collection(dotUsdtCollectionName);
+}
+
+async function tryInit(col) {
+  if (!col) {
+    await initDb();
+  }
+}
+
+async function getKsmUsdtCollection() {
+  await tryInit(ksmUsdtCol);
+  return ksmUsdtCol;
+}
+
+async function getDotUsdtCollection() {
+  await tryInit(dotUsdtCol);
+  return dotUsdtCol;
+}
+
+module.exports = {
+  initDb,
+  getKsmUsdtCollection,
+  getDotUsdtCollection,
+};
