+++ packages/server/src/websocket/constants.js
@@ -1,9 +1,11 @@
 const chainStatusRoom = "CHAIN_STATUS_ROOM";
 const overviewRoom = "OVERVIEW_ROOM";
+const overviewRoomV2 = "OVERVIEW_ROOM_V2";
 
 module.exports = {
   chainStatusRoom,
   FEED_INTERVAL: 6000,
   overviewRoom,
+  overviewRoomV2,
   OVERVIEW_FEED_INTERVAL: 30 * 1000, // 30 seconds
 };
