+++ packages/server/src/websocket/overview/common/constants.js
@@ -1,0 +1,13 @@
+const bountyStatuses = [
+  "Proposed",
+  "Approved",
+  "Funded",
+  "CuratorProposed",
+  "Active",
+  "PendingPayout",
+  "Claimed",
+];
+
+module.exports = {
+  bountyStatuses,
+};
