+++ packages/server/src/websocket/overview/index.js
@@ -1,0 +1,4 @@
+module.exports = {
+  ...require("./v1"),
+  ...require("./v2"),
+};
