+++ packages/server/src/websocket/store.js
@@ -2,10 +2,12 @@ const data = {
   kusama: {
     scanHeight: 0,
     overview: null,
+    overviewV2: null,
   },
   polkadot: {
     scanHeight: 0,
     overview: null,
+    overviewV2: null,
   },
 };
 
@@ -25,9 +27,19 @@ function getOverview(chain) {
   return data[chain]?.overview;
 }
 
+function setOverviewV2(chain, arg) {
+  data[chain].overviewV2 = arg;
+}
+
+function getOverviewV2(chain) {
+  return data[chain]?.overviewV2;
+}
+
 module.exports = {
   setScanHeight,
   getScanHeight,
   setOverview,
   getOverview,
+  setOverviewV2,
+  getOverviewV2,
 };
